+++ src/main/java/gift/dto/OrderDTO.java
@@ -1,18 +1,18 @@
 package gift.dto;
 
 public class OrderDTO {
-    private Long OptionId;
+    private Long optionId;
     private int quantity;
     private String message;
 
-    public OrderDTO(Long OptionId, int quantity, String message) {
-        this.OptionId = OptionId;
+    public OrderDTO(Long optionId, int quantity, String message) {
+        this.optionId = optionId;
         this.quantity = quantity;
         this.message = message;
     }
 
     public Long getOptionId() {
-        return OptionId;
+        return optionId;
     }
 
     public int getQuantity() {
